requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asn1crypto==1.5.1
boto3==1.42.89
botocore==1.42.89
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
fastapi==0.135.3
filelock==3.28.0
h11==0.16.0
idna==3.11
jmespath==1.1.0
numpy==2.4.4
packaging==26.0
pandas==3.0.2
platformdirs==4.9.6
pycparser==3.0
pydantic==2.13.1
pydantic_core==2.46.1
PyJWT==2.12.1
pyOpenSSL==26.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.33.1
s3transfer==0.16.0
six==1.17.0
snowflake-connector-python==4.4.0
sortedcontainers==2.4.0
starlette==1.0.0
tomlkit==0.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.44.0
